发现效果不如之前啊，estimate081-4预测误差不足1cm，难道是ros通信的原因？因为那几个我是在nx上play的，马上试一下在pc上

+++ src/ukf_real.py
@@ -7,7 +7,7 @@ from filterpy.common import Q_discrete_white_noise
 from filterpy.kalman import MerweScaledSigmaPoints
 
 FIRSTFRAME, SECONDFRAME = True, True
-FirstPose, pret, dt, ukf_filter = np.array([0,0,0], dtype=np.float), 0, 0, 0
+FirstPose, pret, dt, ukf_filter = np.array([0.0,0.0,0.0]), 0, 0, 0
 
 def ballPoseCallback(msg, ball_xv_pub):
 	global FIRSTFRAME, SECONDFRAME, FirstPose, pret, dt, ukf_filter
